Refactor ActionService into install/get pattern
