+++ examples/analytics.amp.html
@@ -48,7 +48,7 @@
 <script type="application/json">
 {
   "vars": {
-    "account": "UA-XXXX-Y"
+    "account": "UA-YYYY-Y"
   },
   "triggers": {
     "default pageview": {
